Refactoring, sb preview config
