+++ client/.storybook/preview.js
@@ -1,4 +1,19 @@
+import GlobalStyle from '../src/components/GlobalStyle'
+import { INITIAL_VIEWPORTS } from '@storybook/addon-viewport'
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
+  viewport: {
+    viewports: INITIAL_VIEWPORTS,
+    defaultViewport: 'iphone6',
+  },
 }
+
+export const decorators =[
+  Story => (
+    <>
+      <GlobalStyle />
+      <Story />
+    </>
+  ),
+]
