+++ client/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import App from './components/App/App'
 
 ReactDOM.render(
   <React.StrictMode>
